+++ lib/view/formulier/Formulier.class.php
@@ -5,7 +5,12 @@ use CsrDelft\Orm\Entity\T;
 
 require_once 'view/View.interface.php';
 require_once 'view/Validator.interface.php';
-require_once 'view/formulier/FormElement.abstract.php';
+require_once 'view/formulier/elementen/FormElement.interface.php';
+require_once 'view/formulier/elementen/HtmlComment.class.php';
+require_once 'view/formulier/elementen/HtmlBbComment.class.php';
+require_once 'view/formulier/elementen/Subkopje.class.php';
+require_once 'view/formulier/elementen/CollapsableSubkopje.class.php';
+require_once 'view/formulier/elementen/FieldSet.class.php';
 require_once 'view/formulier/InvoerVelden.class.php';
 require_once 'view/formulier/GetalVelden.class.php';
 require_once 'view/formulier/KeuzeVelden.class.php';
