+++ lib/view/formulier/elementen/Subkopje.class.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * Subkopje.class.php
+ *
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ */
+class Subkopje extends HtmlComment {
+
+	public $h = 3;
+
+	public function getHtml() {
+		return '<h' . $this->h . ' class="' . get_class($this) . '">' . $this->comment . '</h' . $this->h . '>';
+	}
+
+}
